requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
python-decouple>=3.8,<4.0
PyJWT[crypto]>=2.8,<3.0
requests>=2.31,<3.0
psycopg2-binary>=2.9,<3.0
openpyxl>=3.1,<4.0
pypdf>=5.0,<6.0